Improve and conslidate cinder node build inputs
This moves all of the input values used to build the cinder image into the internal/cinder/config/constants package. This ensures that updating versions that go into the cinder image will only need to take place there.

This also changes the way the Kubernetes version is set. Rather than being hardcoded in a const, it is determined when creating a new node by reading the /cinder/kubernetes_version file.

The Kind base image has also been updated and a fix has been applied to account for changes in the image. The newer versions of the image include the kubelet systemd files which means that the /etc/default/kubelet file was already created.